requirements.txt
torch
transformers
diffusers
pandas
requests
Flask 
Pillow
boto3
ffmpeg-python
moviepy
nltk
tensorflow
keras
Pillow
python-dotenv
pydub
celery
streamlit
customtkinter
tk
